+++ src/main/java/de/maxhenkel/easyvillagers/OutputItemHandler.java
@@ -1,0 +1,19 @@
+package de.maxhenkel.easyvillagers;
+
+import net.minecraft.core.NonNullList;
+import net.minecraft.world.item.ItemStack;
+import net.minecraftforge.items.ItemStackHandler;
+import org.jetbrains.annotations.NotNull;
+
+public class OutputItemHandler extends ItemStackHandler {
+
+    public OutputItemHandler(NonNullList<ItemStack> stacks) {
+        super(stacks);
+    }
+
+    @Override
+    public boolean isItemValid(int slot, @NotNull ItemStack stack) {
+        return false;
+    }
+
+}