+++ src/main/java/de/maxhenkel/easyvillagers/blocks/tileentity/IronFarmTileentity.java
@@ -3,6 +3,7 @@ package de.maxhenkel.easyvillagers.blocks.tileentity;
 import de.maxhenkel.corelib.blockentity.ITickableBlockEntity;
 import de.maxhenkel.corelib.inventory.ItemListInventory;
 import de.maxhenkel.easyvillagers.Main;
+import de.maxhenkel.easyvillagers.OutputItemHandler;
 import de.maxhenkel.easyvillagers.blocks.ModBlocks;
 import de.maxhenkel.easyvillagers.blocks.VillagerBlockBase;
 import de.maxhenkel.easyvillagers.entity.EasyVillagerEntity;
@@ -29,22 +30,25 @@ import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.common.util.LazyOptional;
 import net.minecraftforge.items.CapabilityItemHandler;
 import net.minecraftforge.items.IItemHandlerModifiable;
-import net.minecraftforge.items.ItemStackHandler;
 
 import java.util.Collections;
 import java.util.List;
 
 public class IronFarmTileentity extends VillagerTileentity implements ITickableBlockEntity {
 
-    private static final ResourceLocation GOLEM_LOOT_TABLE = new ResourceLocation("entities/iron_golem");
+    protected static final ResourceLocation GOLEM_LOOT_TABLE = new ResourceLocation("entities/iron_golem");
 
-    private NonNullList<ItemStack> inventory;
+    protected NonNullList<ItemStack> inventory;
 
-    private long timer;
+    protected long timer;
+
+    protected LazyOptional<OutputItemHandler> itemHandler;
 
     public IronFarmTileentity(BlockPos pos, BlockState state) {
         super(ModTileEntities.IRON_FARM, ModBlocks.IRON_FARM.defaultBlockState(), pos, state);
         inventory = NonNullList.withSize(4, ItemStack.EMPTY);
+
+        itemHandler = LazyOptional.of(() -> new OutputItemHandler(inventory));
     }
 
     public long getTimer() {
@@ -74,7 +78,7 @@ public class IronFarmTileentity extends VillagerTileentity implements ITickableB
                 }
             } else if (timer >= getGolemKillTime()) {
                 VillagerBlockBase.playVillagerSound(level, getBlockPos(), SoundEvents.IRON_GOLEM_DEATH);
-                IItemHandlerModifiable itemHandler = getItemHandler();
+                IItemHandlerModifiable itemHandler = this.itemHandler.orElse(null);
                 for (ItemStack drop : getDrops()) {
                     for (int i = 0; i < itemHandler.getSlots(); i++) {
                         drop = itemHandler.insertItem(i, drop, false);
@@ -128,21 +132,12 @@ public class IronFarmTileentity extends VillagerTileentity implements ITickableB
         super.load(compound);
     }
 
-    private IItemHandlerModifiable handler;
-
     @Override
     public <T> LazyOptional<T> getCapability(Capability<T> cap, Direction side) {
         if (!remove && cap == CapabilityItemHandler.ITEM_HANDLER_CAPABILITY) {
-            return LazyOptional.of(this::getItemHandler).cast();
+            return itemHandler.cast();
         }
         return super.getCapability(cap, side);
-    }
-
-    public IItemHandlerModifiable getItemHandler() {
-        if (handler == null) {
-            handler = new ItemStackHandler(inventory);
-        }
-        return handler;
     }
 
     public static int getGolemSpawnTime() {
@@ -151,6 +146,12 @@ public class IronFarmTileentity extends VillagerTileentity implements ITickableB
 
     public static int getGolemKillTime() {
         return getGolemSpawnTime() + 20 * 10;
+    }
+
+    @Override
+    public void setRemoved() {
+        itemHandler.invalidate();
+        super.setRemoved();
     }
 
 }