+++ src/main/java/de/maxhenkel/easyvillagers/blocks/tileentity/VillagerTileentity.java
@@ -10,8 +10,8 @@ import net.minecraft.world.level.block.state.BlockState;
 
 public class VillagerTileentity extends FakeWorldTileentity {
 
-    private ItemStack villager;
-    private EasyVillagerEntity villagerEntity;
+    protected ItemStack villager;
+    protected EasyVillagerEntity villagerEntity;
 
     public VillagerTileentity(BlockEntityType<?> type, BlockState defaultState, BlockPos pos, BlockState state) {
         super(type, defaultState, pos, state);